debug

+++ matrix.c
@@ -593,7 +593,7 @@ int mul_matrix(matrix *result, matrix *mat1, matrix *mat2) {
 
                 // result->data[r][c] = mat1->data[r][i] * dst[c * mat2->rows + i] + result->data[r][c];
             }
-            int p[4];
+            double p[4];
             _mm256_storeu_pd(p, result1);
             double sum = p[0] + p[1] + p[2] + p[3];
             for (int i = mat1->cols/4 * 4; i < mat1->cols; i++) {
